Correções:
- Cpf agora é um long, pois o int aguentava apenas 10 algarismos e um
cpf tem 11
- Correção da análise da busca, agora as classes de cadastro só criam a
instância se o id não for encontrado no repositório
- Correção da busca, agora verifica se o objeto no indice do repositório
é nulo

+++ src/models/usuarios/Administrador.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class Administrador extends Usuario {
 
-    public Administrador(int id, String nome, int cpf, LocalDate dataDeNascimento, String email, String senha) {
+    public Administrador(int id, String nome, long cpf, LocalDate dataDeNascimento, String email, String senha) {
         super(id, nome, cpf, dataDeNascimento, email, senha);
     }
