+++ src/models/usuarios/professor/Professor.java
@@ -10,7 +10,7 @@ public abstract class Professor extends Usuario {
     private int idTurma;
     private Salario salario;
 
-    public Professor(int id, String nome, int cpf, LocalDate dataDeNascimento, String email, String senha, int idTurma) {
+    public Professor(int id, String nome, long cpf, LocalDate dataDeNascimento, String email, String senha, int idTurma) {
         super(id, nome, cpf, dataDeNascimento, email, senha);
         this.idTurma = idTurma;
     }
