+++ src/models/usuarios/professor/ProfessorSalarioFixo.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 public class ProfessorSalarioFixo extends Professor implements Salario{
     private double salario;
 
-    public ProfessorSalarioFixo(int id, String nome, int cpf, LocalDate dataDeNascimento, String email, String senha, int idTurma,double salario) {
+    public ProfessorSalarioFixo(int id, String nome, long cpf, LocalDate dataDeNascimento, String email, String senha, int idTurma,double salario) {
         super(id, nome, cpf, dataDeNascimento, email, senha, idTurma);
         this.salario = salario;
     }
